pop: move programs for managing journal into their own folder
The journal will be managed by denis in the future, so disentangling
the journal code from the pop code will be helpful as a first step.
