+++ denis/Containerfile
@@ -9,9 +9,9 @@ COPY --from=run_at_source . /run-at
 
 RUN make -C /run-at CC='clang -static'
 
-COPY --from=pop_source . /pop
+COPY --from=journal_source . /journal
 
-RUN make -C /pop CC='clang -static' restrict_access
+RUN make -C /journal CC='clang -static' restrict_access
 
 WORKDIR /usr/local/share/denis
 COPY . .
@@ -43,7 +43,7 @@ RUN mkdir -p /var/lib/denis && \
 RUN chown -R 100:100 /var/lib/denis
 
 COPY --from=build /run-at/run-at /usr/local/bin/run-at
-COPY --from=build /pop/restrict_access /usr/local/bin/restrict_access
+COPY --from=build /journal/restrict_access /usr/local/bin/restrict_access
 
 USER 100:100
