+++ journal/.gitignore
@@ -1,0 +1,2 @@
+/init_journal
+/restrict_access
