+++ journal/Makefile
@@ -1,0 +1,19 @@
+CC = clang
+CFLAGS = -std=c2x -Weverything -Wno-unsafe-buffer-usage -Wno-declaration-after-statement -Wno-pre-c2x-compat -D_GNU_SOURCE
+
+ifdef DEBUG
+	CFLAGS += -DDEBUG -Og -g
+endif
+
+.PHONY: all clean
+
+all: init_journal restrict_access
+
+init_journal: init_journal.c types.h
+	$(CC) $(CFLAGS) -o $@ $<
+
+restrict_access: restrict_access.c
+	$(CC) $(CFLAGS) -o $@ $<
+
+clean:
+	-rm init_journal restrict_access
