+++ journal/init_journal.c
