+++ journal/restrict_access.c
