+++ pop/.gitignore
@@ -1,3 +1,1 @@
 /pop3
-/init_journal
-/restrict_access
