+++ pop/Containerfile
@@ -8,12 +8,15 @@ COPY --from=tcp_server_source . /tcp_server
 ARG LISTEN_PORT=995
 RUN make -C /tcp_server CC='clang -static' DEFAULT_PORT=${LISTEN_PORT}
 
+COPY --from=journal_source . /journal
+RUN make -C /journal CC='clang -static'
+
 COPY . /pop
 
 RUN make -C /pop CC='clang -static'
 
 RUN mkdir -p /var/lib/email/journal && \
-	/pop/init_journal /var/lib/email/journal/journal && \
+	/journal/init_journal /var/lib/email/journal/journal && \
 	chown -R 100:100 /var/lib/email/journal && \
 	:
 
@@ -21,8 +24,8 @@ FROM scratch as pop
 
 COPY --from=build /tcp_server/tcp_server /usr/local/bin/tcp_server
 COPY --from=build /pop/pop3 /usr/local/bin/pop3
-COPY --from=build /pop/init_journal /usr/local/bin/init_journal
-COPY --from=build /pop/restrict_access /usr/local/bin/restrict_access
+COPY --from=build /journal/init_journal /usr/local/bin/init_journal
+COPY --from=build /journal/restrict_access /usr/local/bin/restrict_access
 COPY --from=build /var/lib/email/ /var/lib/email/
 
 USER 100:100
